+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [3.6.7](https://github.com/easingthemes/dx-aem-flow/compare/v3.6.6...v3.6.7) (2026-09-16)
+
+
+### Bug Fixes
+
+* fetch-raw-story fails on ADO MCP v2.10.0; harden yaml_val and update-progress ([#198](https://github.com/easingthemes/dx-aem-flow/issues/198)) ([32ef945](https://github.com/easingthemes/dx-aem-flow/commit/32ef9459f72cc7ce2622a5de39e61baa71d8b97a)), closes [#2740509](https://github.com/easingthemes/dx-aem-flow/issues/2740509) [#2612832](https://github.com/easingthemes/dx-aem-flow/issues/2612832) [#199](https://github.com/easingthemes/dx-aem-flow/issues/199) [#2740509](https://github.com/easingthemes/dx-aem-flow/issues/2740509) [#2612832](https://github.com/easingthemes/dx-aem-flow/issues/2612832)
+
+## [3.6.6](https://github.com/easingthemes/dx-aem-flow/compare/v3.6.5...v3.6.6) (2026-09-16)
+
+
+### Bug Fixes
+
+* update fetch-raw-story.js for @azure-devops/mcp v2.9.0 tool consolidation ([#197](https://github.com/easingthemes/dx-aem-flow/issues/197)) ([7eb737a](https://github.com/easingthemes/dx-aem-flow/commit/7eb737a4e4e8771c7a313f6e3d0a9869eeb63909)), closes [#198](https://github.com/easingthemes/dx-aem-flow/issues/198)
+
 ## [3.6.5](https://github.com/easingthemes/dx-aem-flow/compare/v3.6.4...v3.6.5) (2026-09-15)